+++ .env.dev
@@ -1,4 +1,3 @@
-# RUN (docker container)
 PORT=3000
 NODE_ENV=development
 POSTGRES_USER=postgres
@@ -10,14 +9,3 @@ JWT_SECRET=****
 JWT_EXPIRES_IN=*d
 JWT_COOKIE_EXPIRES_IN=*
 
-# RUN (localhost)
-# PORT=3001
-# NODE_ENV=development
-# POSTGRES_USER=postgres
-# POSTGRES_PASSWORD=postgres@admin
-# POSTGRES_PORT=5430
-# POSTGRES_HOST=localhost
-# POSTGRES_DB=hux
-# JWT_SECRET=mySecret
-# JWT_EXPIRES_IN=90d
-# JWT_COOKIE_EXPIRES_IN=90
