+++ src/db/db-pool.ts
@@ -5,9 +5,9 @@ export const connectDataSource = async (): Promise<void> => {
   try {
     await dataSource.initialize();
 
-    logger.info('Data source connection: success ✅');
+    logger.info('DB connection: success ✅');
   } catch (error) {
-    logger.error(`Data source connection: failed ❌ - ${JSON.stringify(error)}`);
+    logger.error(`DB connection failed: ${JSON.stringify(error)}`);
 
     process.exit(1);
   }
