+++ src/db/index.ts
@@ -1,2 +1,4 @@
 export * from './data-source';
 export * from './db-pool';
+export * from './entities';
+export * from './model';
