+++ src/index.ts
@@ -1,6 +1,6 @@
 import { createApp } from './app';
 import { startServer } from './server';
-import { Env } from './interfaces';
+import { Env } from './types';
 
 if (process.env.NODE_ENV !== Env.TEST) {
   const app = createApp();
