--- src/interfaces/type.ts
@@ -1,11 +1,0 @@
-enum Role {
-  USER = 'user',
-  ADMIN = 'admin',
-}
-
-interface ObjectLiteral {
-  [props: string]: any;
-}
-
-
-export { Role, ObjectLiteral };
