+++ src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './response';
 export * from './environment';
-export * from './type';
